pyproject.toml
[project]
name = "color-perception-bench"
version = "0.2.0"
description = "Multi-model benchmark for evaluating color perception in multimodal embeddings"
readme = "README.md"
authors = [
    { name = "Michael Pilosov", email = "consistentbayes@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.9.0",
    "joblib>=1.5.3",
    "matplotlib>=3.10.8",
    "numpy>=1.26.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "questionary>=2.0.0",
    "requests>=2.32.5",
    "rich>=13.0.0",
    "tqdm>=4.67.1",
]

[project.scripts]
color-perception-bench = "color_perception_bench.cli:main"

[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"
